# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Backend Developer Test Documentation",
        "description": "L5 Swagger description",
        "contact": {
            "email": "sandip.sathavara@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1/",
            "description": "APIs List"
        }
    ],
    "paths": {
        "/admin/login": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "",
                "description": "Login an Admin account",
                "operationId": "authenticate",
                "parameters": [
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Admin user emailId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Admin user Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of products",
                "description": "Returns list of products",
                "operationId": "index",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "The End date of the project.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Products",
            "description": "API Endpoints of Projects"
        }
    ],
    "security": [
        []
    ]
}